engine/pyproject.toml
[project]
name = "push-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
